Add files via upload
This is the Task - 1 from the CODSOFT in Python Programming language. The task - 1 is A "TO-DO LIST application" is a useful project that helps users manage and organize their tasks efficiently. This project aims to create a command-line or GUI-bsed application using python, allowing users to create, update, and track their to-do lists